# config/swagger.yaml
openapi: 3.0.3
info:
  title: Orders API
  description: API for managing orders and user authentication
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
  - url: https://51ed-157-38-115-182.ngrok-free.app
    description: ngrok server
paths:
...
  /products:
    get:
      summary: Retrieve all products
      description: Endpoint to get a list of all products.
      operationId: getProducts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response with a list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized, invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An unexpected error occurred"
    post:
      summary: Create a new product
      description: Endpoint to create a new product.
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "pen"
                description:
                  type: string
                  example: "blue pen"
                price:
                  type: number
                  format: float
                  example: 10
                stock:
                  type: integer
                  example: 100
              required:
                - name
                - description
                - price
                - stock
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: "Product created successfully"
        '400':
          description: Bad Request, invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid input data"
        '401':
          description: Unauthorized, invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An unexpected error occurred"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authorization
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Sample Product"
        description:
          type: string
...